backend/requirements.txt
Flask
flask-limiter
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy
sqlalchemy
chromadb
openai-agents
openai-chatkit
fastapi
asyncio
uv
uvicorn 
asgiref
psycopg2-binary==2.9.9
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.1.2

reportlab
tiktoken
anthropic
watchdog
celery
redis
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pytz
waitress
gunicorn
stripe


